moved the get_cookie_value() function in the lib.js source file
